requirements.txt
""" Разработка """
pip             22.0.2
setuptools      59.6.0
python          3.9

""" БД """
psycopg2-binary 2.9.10

""" Конфигурация """
python-dotenv   1.1.1
